Optimize memory usage in localedata backing data structure
Previously, when following an alias or merging a loaded locale with a parent, we
copied both dicts into a new merged dict.  For larger dictionaries, this wasted
a lot of memory.

This commit uses a by-reference merged view datastructure for merging larger
dictionaries, saving that memory.  On loading a relatively large number of
locales, it saves roughly 40% of the memory.  If you also access most of the
keys in the locales, the memory savings go up to approximately 60% over the
unmodified code (in other words, the previous code takes over 400Mb where this
code now takes about 160Mb, if you load every locale and access every key in
every one of the resulting dicts)

The by-reference merging is somewhat slower.  For the operation:
* Load every locale
* Walk every locale's dict, recursively, thrice
the unmodified code takes 206s vs the modified code takes 260s on a MacBook Pro
(Retina, 13-inch, Early 2015)

There is no open ticket for this change.